Implement RaspberryPiPkg (based on MS code base)
RaspberryPiPkg implements supporting code for the Raspberry Pi 2 and 3
and is based upon Microsoft's initial board port. This code has been
reorganized and has been adapted to be built in Linux. Support for building
this firmware in Windows will eventually be reimplemented.

The AcpiTables package for both the Raspberry Pi 2 and 3 has been modified to be buildable
with the standard iasl ACPI compiler.

Everything else at this point is still source-identical to Microsoft's ms-iot tree.